requirements.txt
flask==2.2.3
tensorflow==2.11.0
keras==2.11.0
Pillow==9.0.1
requests==2.28.2
whitenoise==6.4.0
gtts==2.3.1
playsound==1.3.0
numpy==1.24.2
matplotlib==3.7.1
gunicorn==20.1.0
protobuf==3.12.4
